tests/wpt/tests/encrypted-media/encrypted-media-supported-by-permissions-policy.tentative.html

<!DOCTYPE html>
<title>Test that encrypted-media is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://github.com/w3c/encrypted-media/pull/432">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('encrypted-media', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise encrypted-media.');
</script>
